==> src/01_github_filter.ipynb <==
# GitHub Filter

This notebook filter the projects based on GitHub metadata. 

This notebook filter the projects based on the following criteria:

 - developed using one of the top 10 languages from stackoverflow survey
 - had at least 10 tags

The results are stored in the `data/01_github_project_results.json` and we cloned the projects using the `src/02_github_clone.sh`


import json
import os

from util import DATA_PATH

with open(DATA_PATH / '00_github_project_search-results.json', 'r') as infile:
    projects = json.load(infile)
    
print("GitHub Projects: {:,}".format(len(projects)))

# define top languages
top_langs = set(lang.lower() for lang in ['JavaScript', 'Python', 'Java', 'TypeScript', 'C#', 'C++', 'PHP', 'C', 'Go', 'Kotlin'])
top_langs

# Apply filters
def project_criteria(project):
    # fewer potential releases
    if project['tags']['totalCount'] < 10:
        return False
    
    # no language
    if not project['topLanguage']['edges']:
        return False
    
    # unselected language
    if project['topLanguage']['edges'][0]['node']['name'].lower() not in top_langs:
        return False
    
    return True
        
selected_projects = [project for project in filter(project_criteria, projects)]
    
# selected_projects = [project for project in projects
#                      if project['tags']['totalCount'] >= 10
#                  and projects[0]['topLanguage']['edges'][0]['node']['name'].lower() in set(lang.lower() for lang in ['JavaScript', 'Python', 'Java', 'TypeScript', 'C#', 'C++', 'PHP', 'C', 'Go', 'Kotlin'])]

print("Selected projects: {:,}".format(len(selected_projects)))

print("Tags: {:,}".format(sum(map(lambda project: project['tags']['totalCount'], selected_projects))))

print("HEAD commits: {:,}".format(sum(map(lambda project: project['commits']['history']['totalCount'], selected_projects))))

with open(DATA_PATH / '01_github_project_results.json', 'w') as outfile:
    json.dump(selected_projects, outfile)

with open(DATA_PATH / '01_github_project_results.txt', 'w') as outfile:
    outfile.write(
        '\n'.join(
            map(lambda p: p['nameWithOwner'], selected_projects)))
