==> src/10_mine_projects.ipynb <==
# Mine and select projects

This notebook mine the projects previous cloned using releasy

import os
import datetime
import json
import re
import traceback

from multiprocessing import Pool
import pandas as pd

from ipywidgets import IntProgress
from IPython.display import display

import releasy

from util import (
    DATA_PATH,
    REPO_PATH,
    TMP_PATH, 
    CPU,
    CycleType,
    RAPID_RELEASE_LIM,
    TRAD_RELEASE_LIM,
    delta2days,
    is_rapid_release,
    is_trad_release,
)

pd.set_option('display.max_rows', 1000)

with open(DATA_PATH / '01_github_project_results.json', 'r') as infile:
    projects_data = json.loads(infile.read())

print(len(projects_data))

## Mine Releases


def mine(name: str) -> releasy.Project:
    elapsed_time = datetime.datetime.now()
    
    try:
        repo_path = str(REPO_PATH / name)
        project = releasy.Miner(repo_path, name).apply(
            releasy.FinalReleaseMiner(),
            releasy.HistoryCommitMiner(),
            releasy.BaseReleaseMiner(),
            releasy.ContributorMiner(),
            releasy.SemanticReleaseMiner()
        ).mine()

        rapid_releases = [release for release in project.main_releases if is_rapid_release(release) and len(release.commits) >= 10]
        trad_releases = [release for release in project.main_releases if is_trad_release(release) and len(release.commits) >= 10]

        project_data = {
            'project': project.name,
            'prefixes': len(project.releases.prefixes()),
            'prefixes_names': str(" ".join(project.releases.prefixes())),
            'main_releases': len(project.main_releases),
            'rapid_releases': len(rapid_releases),
            'trad_releases': len(trad_releases),
            'patches': len(project.patches),
            'releases': len(project.releases)
        }

        elapsed_time = datetime.datetime.now() - elapsed_time
        project_data['time'] = elapsed_time
    except Exception as err:
        print(f"{name:40} {err=}")
        traceback.print_exception(err)
        project_data = {}
        mreleases = [{}]
    
    return project_data

project_names = list(map(lambda p: p['nameWithOwner'], projects_data))
# project_names = project_names[:20]

with Pool(processes=CPU) as pool:
    processed = 0
    results = []
    progress = IntProgress(min=0, max=len(project_names))
    display(progress)
    for result in pool.imap_unordered(mine, project_names):
        results.append(result)
        progress.value += 1
        
projects = pd.DataFrame(results)
projects.sample(10)

len(projects)

projects.to_csv(DATA_PATH / '10_projects_all.csv')

## Select Releases

selected_projects = projects.copy()

### Remove projects without main releases

These projects may contain releases, but releasy did not extracted any main release. Hence, we need to remove these projects from the analysis

# total = len(selected_projects)
# selected_projects = selected_projects.query('main_releases > 0')
# removed = total - len(selected_projects)
# print(f"Discarded {removed} projects")
# print(f"Kept {len(selected_projects)} projects")

### Remove without few rapid and traditional releases

The projects must have at least one rapid and one traditional release

total = len(selected_projects)
selected_projects = selected_projects.query('rapid_releases >= 2 and trad_releases >= 2')
removed = total - len(selected_projects)
print(f"Discarded {removed} projects")
print(f"Kept {len(selected_projects)} projects")

### Remove projecs with uncommon prefixes

These projects usually represent mono repo, i.e., multiple software projects in the same git repository. Hence, we need to remove these projects from the analysis

def uncommon_prefixes(qnt, names):
    uncommon_prefixes = 0
    if qnt <= 1:
        return uncommon_prefixes
    else:
        names = str(names).split(' ')
        uncommon_prefixes = [name for name in names if not re.match("^((|vv?|rel|l|rel/v|r|release)[/._-]?)$", name, flags=re.IGNORECASE)]
        return len(uncommon_prefixes)
        
selected_projects = selected_projects.assign(
    uncommon_prefixes=selected_projects.apply(lambda row: uncommon_prefixes(row['prefixes'], row['prefixes_names']), axis = 1))

len(selected_projects.query('uncommon_prefixes > 0'))

total = len(selected_projects)
selected_projects = selected_projects.query('uncommon_prefixes == 0')
removed = total - len(selected_projects)
print(f"Discarded {removed} projects")
print(f"Kept {len(selected_projects)} projects")

selected_projects.sample(10)

len(selected_projects)

selected_projects.to_csv(DATA_PATH / '10_projects_selected.csv', index=False)
